apps/requirements.txt
# apps/requirements.txt
kafka-python==2.0.2
requests==2.31.0
prometheus_client==0.20.0